Backend/LocalMVP/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
chromadb==1.3.4
openai==2.8.0
google-genai==1.50.0
